+++ src/main/java/com/asg/shipchandling/salesinvoice/dto/response/RefreshGpProcResponse.java
@@ -4,21 +4,23 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class RefreshGpProcResponse {
     private Boolean success;
     private String message;
-    private Long discountPercent;
-    private Long discountAmt;
-    private Long incentivePercent;
-    private Long incentivePercent2;
-    private Long incentivePercent3;
-    private Long incentiveAmt;
-    private Long incentiveAmt2;
-    private Long incentiveAmt3;
-    private Long totalGpAmt;
-    private Long totalGpPercent;
-    private Long invAmount;
+    private BigDecimal discountPercent;
+    private BigDecimal discountAmt;
+    private BigDecimal incentivePercent;
+    private BigDecimal incentivePercent2;
+    private BigDecimal incentivePercent3;
+    private BigDecimal incentiveAmt;
+    private BigDecimal incentiveAmt2;
+    private BigDecimal incentiveAmt3;
+    private BigDecimal totalGpAmt;
+    private BigDecimal totalGpPercent;
+    private BigDecimal invAmount;
 }
